Overhauled layer system
Turns out, using layer polymorphism is a bad approach for compatibility with my GPU code. It seems dynamic casts are not supported in CUDA. It's a bit of a shame because it would've made the CPU code modular and easily expandable. This commit removes my polymorphism approach and replaces it with a simpler Layer struct with a layerType member.